+++ uploader/handlers/push_dataset_events.py
@@ -94,14 +94,17 @@ def handler(event, context):
     )
     target_s3_path_raw = generate_s3_path(dataset, edition["Id"], "raw", absolute=True)
 
-    log_add(target_s3_path=target_s3_path)
+    log_add(
+        target_s3_path_processed=target_s3_path_processed,
+        target_s3_path_raw=target_s3_path_raw,
+    )
 
     # Write the raw input data
-    s3.put_object(
-        Body=json.dumps(body["events"]),
-        Bucket=os.environ["BUCKET"],
-        Key=f"{target_s3_path_raw}/data.json",
-    )
+    # s3.put_object(
+    #     Body=json.dumps(body["events"]),
+    #     Bucket=os.environ["BUCKET"],
+    #     Key=f"{target_s3_path_raw}/data.json",
+    # )
 
     # Write merged data to both the new edition and to `latest`
     for path in target_s3_path_processed, source_s3_path:
